Clean-up types
