+++ packages/rainbow-sprinkles/src/createRainbowSprinkles.ts
@@ -1,31 +1,18 @@
-import type { RuntimeFnReturn } from './types';
-import type { ChildSprinkles, SprinkleProperties } from './exp';
+import type {
+  RuntimeFnReturn,
+  DefinePropertiesReturn,
+  SprinklesProps,
+} from './types';
 import { addFunctionSerializer } from '@vanilla-extract/css/functionSerializer';
 import { createRuntimeFn } from './createRuntimeFn';
 
-export type SprinklesProps<Args extends ReadonlyArray<any>> = Args extends [
-  infer L,
-  ...infer R,
-]
-  ? (L extends BaseReturn ? ChildSprinkles<L['config']> : never) &
-      SprinklesProps<R>
-  : {};
-
-export type BaseReturn = {
-  config: SprinkleProperties;
-  conditions: {
-    defaultCondition: string;
-    conditionNames: string[];
-  };
-};
-
-export type SprinklesFn<Args extends ReadonlyArray<BaseReturn>> = (
+export type SprinklesFn<Args extends ReadonlyArray<DefinePropertiesReturn>> = (
   props: SprinklesProps<Args>,
 ) => RuntimeFnReturn;
 
-export function createRainbowSprinkles<Args extends ReadonlyArray<BaseReturn>>(
-  ...args: Args
-): SprinklesFn<Args> {
+export function createRainbowSprinkles<
+  Args extends ReadonlyArray<DefinePropertiesReturn>,
+>(...args: Args): SprinklesFn<Args> {
   const cssConfig = Object.assign({}, ...args.map((a) => a.config));
   const properties = Object.keys(cssConfig);
